pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["clean-pii"]

[project]
name = "clean-pii"
version = "0.1.0"
authors = [
    { name="Denis Musinguzi", email="musinguzidenis97@gmail.com" },
]
description = "A repository for object detection of PII for removal"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "ultralytics",
    "torch",
    "opencv-python",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "mypy",
    "pre-commit",
    "pandas-stubs",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]



[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]